# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CommerceOS API
  version: 0.1.0
  description: Authentication, public catalogue, wishlist, and cart contract.
servers:
  - url: /api/v1
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a customer and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201': { $ref: '#/components/responses/Session' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the refresh cookie and issue a new access token
      security: [{ refreshCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke this refresh-token family
      security: [{ bearerAuth: [], refreshCookie: [] }]
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset; always responds identically for known and unknown accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              additionalProperties: false
              properties:
                email: { type: string, format: email }
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message: { type: string }
        '429': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Consume a one-time reset token and revoke all refresh sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              additionalProperties: false
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '204': { description: Password reset }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Read the authenticated account
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/User' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /me:
    patch:
      tags: [Account]
      summary: Update the authenticated profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: false
              properties:
                firstName: { type: string }
                lastName: { type: string }
                phone: { type: string }
      responses:
        '200': { $ref: '#/components/responses/User' }
        default: { $ref: '#/components/responses/Error' }
  /me/change-password:
    post:
      tags: [Account]
      summary: Change password and revoke refresh sessions
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              additionalProperties: false
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '204': { description: Password changed }
        default: { $ref: '#/components/responses/Error' }
  /me/addresses:
    get:
      tags: [Account]
      summary: List own addresses
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/AddressList' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Account]
      summary: Add an address (maximum ten)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
      responses:
        '201': { $ref: '#/components/responses/Address' }
        default: { $ref: '#/components/responses/Error' }
  /me/addresses/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [Account]
      summary: Update an owned address
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressUpdate' }
      responses:
        '200': { $ref: '#/components/responses/Address' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Account]
      summary: Delete an owned address
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Address deleted }
        default: { $ref: '#/components/responses/Error' }
  /admin/auth/check:
    get:
      tags: [Auth]
      summary: Verify admin access
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Admin access granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      allowed: { type: boolean, const: true }
        '403': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /system/echo:
    get:
      summary: Validate and echo a message
      parameters:
        - in: query
          name: message
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Echoed message
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [message]
                    properties:
                      message:
                        type: string
        '422':
          $ref: '#/components/responses/ValidationError'
  /health:
    servers:
      - url: /
    get:
      summary: MongoDB-backed health check
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [status]
                    properties:
                      status:
                        const: ok
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/public:
    get:
      tags: [Catalogue]
      summary: Public store profile, currency, shipping, theme and feature flags
      responses:
        '200': { $ref: '#/components/responses/Settings' }
        default: { $ref: '#/components/responses/Error' }
  /categories:
    get:
      tags: [Catalogue]
      summary: Active category tree and visible product counts
      responses:
        '200': { $ref: '#/components/responses/Categories' }
        default: { $ref: '#/components/responses/Error' }
  /categories/{slug}:
    get:
      tags: [Catalogue]
      summary: Active category, children and breadcrumbs
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Category' }
        '404': { $ref: '#/components/responses/Error' }
  /products:
    get:
      tags: [Catalogue]
      summary: Paginated, filtered product search
      description: Option filters use options[Color]=Graphite. sort=relevance requires q.
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: minPrice, in: query, schema: { type: integer, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: integer, minimum: 0 } }
        - { name: inStock, in: query, schema: { type: boolean } }
        - { name: rating, in: query, schema: { type: number, minimum: 0, maximum: 5 } }
        - { name: featured, in: query, schema: { type: boolean } }
        - { name: tags, in: query, schema: { type: string } }
        - { name: 'options[Color]', in: query, schema: { type: string } }
        - { name: includeFacets, in: query, schema: { type: boolean } }
        - {
            name: sort,
            in: query,
            schema:
              {
                type: string,
                enum: [relevance, -createdAt, price, -price, -soldCount, -ratingAverage, name],
              },
          }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { $ref: '#/components/responses/Products' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /products/suggest:
    get:
      tags: [Catalogue]
      summary: Up to six product and category search suggestions
      parameters: [{ name: q, in: query, required: true, schema: { type: string, minLength: 2 } }]
      responses:
        '200': { $ref: '#/components/responses/Suggestions' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /products/{slug}:
    get:
      tags: [Catalogue]
      summary: Visible product with active variants, gallery and specifications
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Product' }
        '404': { $ref: '#/components/responses/Error' }
  /products/{slug}/related:
    get:
      tags: [Catalogue]
      summary: Up to eight related visible products
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Products' }
        '404': { $ref: '#/components/responses/Error' }
  /wishlist:
    get:
      tags: [Wishlist]
      summary: Paginated saved products
      security: [{ bearerAuth: [] }]
      parameters:
        [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { $ref: '#/components/responses/Products' }
        '401': { $ref: '#/components/responses/Error' }
  /wishlist/{productId}:
    parameters: [{ $ref: '#/components/parameters/ProductId' }]
    post:
      tags: [Wishlist]
      summary: Save a visible product, idempotently
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Product saved }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Wishlist]
      summary: Remove a saved product
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Product removed }
        '401': { $ref: '#/components/responses/Error' }
  /cart/price:
    post:
      tags: [Cart]
      summary: Quote guest items against current catalogue prices and stock
      requestBody: { $ref: '#/components/requestBodies/CartItems' }
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /cart:
    get:
      tags: [Cart]
      summary: Read and reprice own cart
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '401': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Cart]
      summary: Empty own cart
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Cart cleared }
  /cart/items:
    post:
      tags: [Cart]
      summary: Add an item, enforcing quantity and stock
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/CartItem' }
      responses:
        '201': { $ref: '#/components/responses/Cart' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /cart/items/{itemId}:
    parameters: [{ $ref: '#/components/parameters/ItemId' }]
    patch:
      tags: [Cart]
      summary: Set item quantity
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [quantity],
                    properties: { quantity: { type: integer, minimum: 1, maximum: 10 } },
                  },
              },
          }
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Cart]
      summary: Remove an item
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Item removed }
        '404': { $ref: '#/components/responses/Error' }
  /cart/merge:
    post:
      tags: [Cart]
      summary: Merge guest lines with own cart, capped by stock and ten units
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/CartItems' }
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /cart/discount:
    post:
      tags: [Cart]
      summary: Apply one discount code
      security: [{ bearerAuth: [] }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { type: object, required: [code], properties: { code: { type: string } } },
                },
            },
        }
      responses:
        '200': { $ref: '#/components/responses/Cart' }
        '422': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Cart]
      summary: Remove discount code
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Discount removed }
components:
  parameters:
    Slug: { name: slug, in: path, required: true, schema: { type: string } }
    ProductId:
      {
        name: productId,
        in: path,
        required: true,
        schema: { type: string, pattern: '^[a-fA-F0-9]{24}$' },
      }
    ItemId:
      {
        name: itemId,
        in: path,
        required: true,
        schema: { type: string, pattern: '^[a-fA-F0-9]{24}$' },
      }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 12 } }
  requestBodies:
    CartItem:
      {
        required: true,
        content: { application/json: { schema: { $ref: '#/components/schemas/CartItemInput' } } },
      }
    CartItems:
      {
        required: true,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [items],
                    properties:
                      {
                        items:
                          { type: array, items: { $ref: '#/components/schemas/CartItemInput' } },
                      },
                  },
              },
          },
      }
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: refreshToken }
  schemas:
    CartItemInput:
      type: object
      required: [productId, variantId, quantity]
      additionalProperties: false
      properties:
        productId: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
        variantId: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
        quantity: { type: integer, minimum: 1, maximum: 10 }
    CartLine:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        variantId: { type: string }
        quantity: { type: integer }
        name: { type: string }
        slug: { type: string }
        sku: { type: string }
        variantLabel: { type: string }
        image: { type: string }
        unitPrice: { type: integer, description: Current server price in minor units }
        lineTotal: { type: integer }
        available: { type: integer }
        unavailable: { type: boolean }
        priceChanged: { type: boolean }
    CartQuote:
      type: object
      properties:
        lines: { type: array, items: { $ref: '#/components/schemas/CartLine' } }
        discountCode: { type: [string, 'null'] }
        discountError: { type: [object, 'null'] }
        currency: { type: string }
        shippingMethod: { type: [object, 'null'] }
        totals:
          type: object
          properties:
            subtotal: { type: integer }
            discount: { type: integer }
            shipping: { type: integer }
            tax: { type: integer }
            total: { type: integer }
        canCheckout: { type: boolean }
    ProductSummary:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        primaryImage: { type: [object, 'null'] }
        priceFrom: { type: integer }
        compareAtFrom: { type: [integer, 'null'] }
        inStock: { type: boolean }
        ratingAverage: { type: number }
        ratingCount: { type: integer }
        category: { type: object }
    ProductDetail:
      allOf:
        - { $ref: '#/components/schemas/ProductSummary' }
        - type: object
          properties:
            description: { type: string }
            brand: { type: string }
            images: { type: array, items: { type: object } }
            specs: { type: array, items: { type: object } }
            optionDefinitions: { type: array, items: { type: object } }
            variants: { type: array, items: { type: object } }
            seo: { type: object }
    Category:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        image: { type: string }
        parentId: { type: [string, 'null'] }
        productCount: { type: integer }
        children: { type: array, items: { type: object } }
    PublicSettings:
      type: object
      properties:
        store: { type: object }
        currency: { type: object }
        tax: { type: object }
        shipping: { type: object }
        inventory: { type: object }
        theme: { type: object }
        features: { type: object }
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        firstName: { type: string }
        lastName: { type: string }
        role: { description: Ignored; every public registration creates a customer }
    LoginInput:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AddressInput:
      type: object
      required: [label, fullName, line1, city, region, postalCode, country, phone]
      additionalProperties: false
      properties:
        label: { type: string }
        fullName: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        region: { type: string }
        postalCode: { type: string }
        country: { type: string, minLength: 2, maxLength: 2 }
        phone: { type: string }
        isDefault: { type: boolean }
    AddressUpdate:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        label: { type: string }
        fullName: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        region: { type: string }
        postalCode: { type: string }
        country: { type: string, minLength: 2, maxLength: 2 }
        phone: { type: string }
        isDefault: { type: boolean }
    Address:
      type: object
      required: [id, label, fullName, line1, city, region, postalCode, country, phone, isDefault]
      properties:
        id: { type: string }
        label: { type: string }
        fullName: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        region: { type: string }
        postalCode: { type: string }
        country: { type: string }
        phone: { type: string }
        isDefault: { type: boolean }
    User:
      type: object
      required: [id, email, firstName, lastName, role, status]
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        role: { type: string, enum: [customer, admin] }
        status: { type: string, enum: [active, disabled] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
  responses:
    Cart:
      description: Current cart quote
      content:
        {
          application/json:
            {
              schema:
                { type: object, properties: { data: { $ref: '#/components/schemas/CartQuote' } } },
            },
        }
    Products:
      description: Visible products with pagination metadata for list operations
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties:
                    {
                      data: { type: array, items: { $ref: '#/components/schemas/ProductSummary' } },
                      meta: { type: object },
                    },
                },
            },
        }
    Product:
      description: Visible product detail
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties: { data: { $ref: '#/components/schemas/ProductDetail' } },
                },
            },
        }
    Categories:
      description: Active categories
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties:
                    { data: { type: array, items: { $ref: '#/components/schemas/Category' } } },
                },
            },
        }
    Category:
      description: Active category detail
      content:
        {
          application/json:
            {
              schema:
                { type: object, properties: { data: { $ref: '#/components/schemas/Category' } } },
            },
        }
    Settings:
      description: Public store settings
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties: { data: { $ref: '#/components/schemas/PublicSettings' } },
                },
            },
        }
    Suggestions:
      description: Product and category suggestions
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties:
                    {
                      data:
                        {
                          type: array,
                          items:
                            {
                              type: object,
                              properties:
                                {
                                  type: { type: string },
                                  name: { type: string },
                                  slug: { type: string },
                                },
                            },
                        },
                    },
                },
            },
        }
    Error:
      description: Request rejected
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    User:
      description: Authenticated account
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: '#/components/schemas/User' }
    Address:
      description: Saved address
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: '#/components/schemas/Address' }
    AddressList:
      description: Saved addresses
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items: { $ref: '#/components/schemas/Address' }
    Session:
      description: Access token in response; refresh token in HttpOnly cookie
      headers:
        Set-Cookie:
          schema: { type: string }
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  accessToken: { type: string }
                  user: { $ref: '#/components/schemas/User' }
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Unexpected failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'